requirements.txt
mysql-connector-python
Jinja2==3.1.6
uvicorn==0.34.0
fastapi==0.115.12
python-dotenv==1.1.0
python-multipart==0.0.20
